requirements.txt
# OCR engines
# Please ensure you have the correct CUDA version installed for paddlepaddle-gpu.
# For CPU-only, use 'paddlepaddle'.
paddleocr

# Vision Transformers
# torch is a dependency for transformers and paddleocr.
torch
transformers>=4.20.0

# LLM integration
openai>=1.0.0

# UI
streamlit>=1.22.0

# Utils
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0

# Evaluation
rapidfuzz

# Data handling
pandas>=1.3.0
